% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_poll_boxplot.R
\name{eg_poll_boxplot}
\alias{eg_poll_boxplot}
\title{Box Plot of Poll}
\usage{
eg_poll_boxplot(poll, abbr = TRUE)
}
\arguments{
\item{poll}{a data.frame with c('Date','value','variable','fullname')}

\item{abbr}{Should names be abbreviated (for legend legibility)}
}
\value{
a ggplot object
}
\description{
This plot function plots a poll over a timeline.
}
