% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_poll_timeline.R
\name{eg_poll_timeline}
\alias{eg_poll_timeline}
\title{Polls on Timeline}
\usage{
eg_poll_timeline(poll, abbr = TRUE)
}
\arguments{
\item{poll}{a data.frame with c('Date','value','variable','fullname')}

\item{abbr}{Should names be abbreviated (for legend legibility)}
}
\value{
a ggplot or a plotly object
}
\description{
This plot function plots a poll over a timeline.
}
