% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eg_poll_validate.R
\name{eg_poll_validate}
\alias{eg_poll_validate}
\title{Validate Poll Dataframe}
\usage{
eg_poll_validate(poll)
}
\arguments{
\item{poll}{a data.frame to validate}
}
\value{
nothing, but stops on error
}
\description{
Validate Poll Dataframe
}
