% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egbarplotcount.R
\name{egbarplotcount}
\alias{egbarplotcount}
\title{Return a ggplot barplot with configuration by default for Ethergeist.}
\usage{
egbarplotcount(data, xCol, fillCol, xlab = "", ylab = "", legendlab = "",
  coord_flip = FALSE)
}
\description{
Plot based on ggplot2 package.
}
\author{
Pablo Pagnone
}
