% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egdataseries.R
\name{egdataseries}
\alias{egdataseries}
\title{Return a ggplot timeseries with configuration by default for Ethergeist.}
\usage{
egdataseries(df, colDate, colFill, ylab = "Date", xlab = "Count",
  legendlab = "")
}
\arguments{
\item{df}{- data.frame to plot}

\item{colFill}{- Column to use as fill and group.}

\item{labcolor}{- Lab for color.}

\item{labx}{- Lab for X}

\item{laby}{- Lab for Y}
}
\description{
Plot based on ggplot2 package.
}
\author{
Pablo Pagnone
}
