% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egheatmap.R
\name{egheatmap}
\alias{egheatmap}
\title{Return a ggplot heatmap with configuration by default for Ethergeist.}
\usage{
egheatmap(data, colx, coly, colfill, xlab = "", ylab = "", legendlab = "")
}
\arguments{
\item{data-}{data.frame to plot}

\item{colFill}{- Column to use as fill and group.}
}
\description{
Plot based on ggplot2 package.
}
\author{
Pablo Pagnone
}
