% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emotion_correlation.R
\name{get_emotion_correlation}
\alias{get_emotion_correlation}
\title{Correlation Between Emotions}
\usage{
get_emotion_correlation(df)
}
\arguments{
\item{df}{a data.frame from syuzhet::get_nrc_sentiment()}
}
\description{
Calculates the correlation between emotions present in the df
}
