% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emotion_summary.R
\name{get_emotion_summary}
\alias{get_emotion_summary}
\title{Get Summary of Emitions in a syuzhet df}
\usage{
get_emotion_summary(df, grouping.var = NULL, fun = "sum")
}
\arguments{
\item{grouping.var}{}
}
\description{
Get Summary of Emitions in a syuzhet df
}
