% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emotion_tally.R
\name{get_emotion_tally}
\alias{get_emotion_tally}
\title{Tally of Emotions}
\usage{
get_emotion_tally(df, grouping.var = NULL)
}
\value{
a data.frame with variable and Value
}
\description{
Tally of Emotions
}
