% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrc_dimensions.R
\name{get_nrc_dimensions}
\alias{get_nrc_dimensions}
\title{Get NRC Emotions}
\usage{
get_nrc_dimensions(language = c("english", "spanish"))
}
\arguments{
\item{language}{}
}
\value{
a labeled character vector of the emotions used in the NRC dictionary
}
\description{
Get NRC Emotions
}
