% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_emotion_cols.R
\name{has_emotion_cols}
\alias{has_emotion_cols}
\title{Validate Syuzhet data.frame}
\usage{
has_emotion_cols(df)
}
\arguments{
\item{df}{}
}
\value{
TRUE if it does FALSE if not.
}
\description{
Checks that the data.frame passed to it is indeed the result of a syuzhet analysis.
Checks to see if the correct columns exist in the data.frame.
}
