% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syu_circular.R
\name{syu_circular}
\alias{syu_circular}
\title{Circular Plot of Emotions}
\usage{
syu_circular(df, language = "english", grouping.var = NULL)
}
\arguments{
\item{language}{english or spanish}
}
\value{
a ggplot
}
\description{
Circular Plot of Emotions
}
