% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syu_corrplot.R
\name{syu_corrplot}
\alias{syu_corrplot}
\title{Correlation Plot}
\usage{
syu_corrplot(df, ...)
}
\arguments{
\item{df}{output from get_nrc_sentiment}

\item{...}{passed to ggcorrplot}
}
\value{
a ggcorrplot
}
\description{
Correlation Plot
}
