% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syu_radar.R
\name{syu_radar}
\alias{syu_radar}
\title{Circular Plot of Emotions}
\usage{
syu_radar(df, fun = "mean", grouping.var = NULL, ...)
}
\arguments{
\item{...}{other parameters passed to eg_radar_plot}

\item{language}{english or spanish}
}
\value{
a ggplot
}
\description{
Circular Plot of Emotions
}
